More automation and docs

+++ res/libsodium/buildlibsodium.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
+cd "$(dirname "$0")/../.."
+set -e
 
 # First thing to do is see if libsodium.a exists in the res folder. If it does, then there's nothing to do
 if [ -f res/libsodium.a ]; then
     rm res/libsodium.a
-    rm -r res/libsodium/libsodium-1.0.21
+    rm -rf res/libsodium/libsodium-1.0.21
 fi
 
 echo "Building libsodium"
@@ -29,5 +31,12 @@ else
 fi
 cd ..
 
-# copy the library to the parents's res/ folder
-cp libsodium-1.0.21/src/libsodium/.libs/libsodium.a ../
+# copy the library to res/ folder
+if [ -f libsodium-1.0.21/src/libsodium/.libs/libsodium.a ]; then
+    cp libsodium-1.0.21/src/libsodium/.libs/libsodium.a ../
+elif [ -f libsodium-1.0.21/.libs/libsodium.a ]; then
+    cp libsodium-1.0.21/.libs/libsodium.a ../
+else
+    echo "ERROR: libsodium.a not found after build"
+    exit 1
+fi
